Create experimental branch
Adds validation for user input fields when creating or updating Employee information. Provides temporal validation for new shifts added to prevent shifts with non-positive durations (such as a shift that starts today and ends yesterday or a shift that starts and ends on the same day at the same time).